requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
APScheduler==3.11.2
attrs==26.1.0
autopep8==2.3.2
bcrypt==5.0.0
beanie==1.28.0
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cryptography==46.0.6
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
distro==1.9.0
dnspython==1.16.0
ecdsa==0.19.2
eventregistry==9.1
fastapi==0.135.2
filelock==3.25.2
flake8==7.3.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
geopandas==1.1.3
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-genai==1.68.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
lazy-model==0.2.0
librt==0.8.1
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
motor==3.6.0
mpmath==1.3.0
multidict==6.7.1
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.10.0
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
osmnx==2.1.0
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyflakes==3.4.0
PyJWT==2.12.1
pymongo==4.9.2
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
pyright==1.1.408
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
triton==3.6.0
twilio==9.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
ultralytics==8.4.30
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
yarl==1.23.0
